+++ src/rpc_client/mod.rs
@@ -6,6 +6,7 @@ pub mod beacon_ops;
 pub mod chain_ops;
 pub mod common_ops;
 pub mod eth_ops;
+pub mod gas_ops;
 pub mod mpool_ops;
 pub mod net_ops;
 pub mod node_ops;